--- tv-app/effecticeTV/app/src/main/java/band/effective/office/tv/network/leader/models/EventInfo/Hall.kt
@@ -1,15 +1,0 @@
-package band.effective.office.tv.network.leader.models.EventInfo
-
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-data class Hall(
-    val capacity: Int,
-    val id: Int,
-    val name: String,
-    val photos: List<Photo>,
-    val preparePeriod: Int,
-    val square: String,
-    val tags: List<String>,
-    val type: String
-)
