--- tv-app/effecticeTV/app/src/main/java/band/effective/office/tv/network/leader/models/EventInfo/ParticipantsItem.kt
@@ -1,10 +1,0 @@
-package band.effective.office.tv.network.leader.models.EventInfo
-
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-data class ParticipantsItem (
-    val id: Int,
-    val name: String,
-    val photo: String?
-)
