--- tv-app/effecticeTV/app/src/main/java/band/effective/office/tv/network/leader/models/EventInfo/PhotoX.kt
@@ -1,9 +1,0 @@
-package band.effective.office.tv.network.leader.models.EventInfo
-
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-data class PhotoX(
-    val full: String,
-    val thumb: ThumbX
-)
