--- tv-app/effecticeTV/app/src/main/java/band/effective/office/tv/network/leader/models/EventInfo/Stat.kt
@@ -1,15 +1,0 @@
-package band.effective.office.tv.network.leader.models.EventInfo
-
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-data class Stat(
-    val active_participants: Int,
-    val moderationTime: String,
-    val monthEventCount: Int,
-    val ntiPercent: Double,
-    val participantAverage: String,
-    val participantCount: Int,
-    val regionScope: Int,
-    val uniqueUsers: Int
-)
