--- tv-app/effecticeTV/app/src/main/java/band/effective/office/tv/network/leader/models/EventInfo/Timezone.kt
@@ -1,9 +1,0 @@
-package band.effective.office.tv.network.leader.models.EventInfo
-
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-data class Timezone(
-    val minutes: Int,
-    val value: String
-)
