--- tv-app/effecticeTV/app/src/main/java/band/effective/office/tv/network/leader/models/SearchEvent/TypeX.kt
@@ -1,9 +1,0 @@
-package band.effective.office.tv.network.leader.models.SearchEvent
-
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-data class TypeX(
-    val id: Int,
-    val name: String
-)
