+++ tv-app/effecticeTV/app/src/main/java/band/effective/office/tv/network/leader/models/TimezoneX.kt
@@ -1,4 +1,4 @@
-package band.effective.office.tv.network.leader.models.SearchEvent
+package band.effective.office.tv.network.leader.models
 
 import com.squareup.moshi.JsonClass
