+++ tv-app/effecticeTV/app/src/main/java/band/effective/office/tv/network/leader/models/TypeX.kt
@@ -1,9 +1,9 @@
-package band.effective.office.tv.network.leader.models.EventInfo
+package band.effective.office.tv.network.leader.models
 
 import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
-data class Type(
+data class TypeX(
     val id: Int,
     val name: String
 )
